add kvmi support
Add the new 'introspection' object with two required options:
 - chardev: the socket connected to the introspection tool
 - key: the key to validate the introspection tool
Add the 'cid' option to chardev/socket object (vsock).
Add the 'fd' class property to chardev/socket.
Add the 'introspection' option to accel/kvm.

The introspector is started with:
	-device vhost-vsock-pci,id=vhost-vsock-pci0,guest-cid=3

The introspected VM is started with:
	-chardev socket,id=chardev0,cid=3,port=1234 \
	-object secret,id=key0,data=some \
	-object introspection,id=kvmi,chardev=chardev0,key=key0 \
	-accel kvm,introspection=kvmi

The 'secret' object is not used yet, but required :)

Signed-off-by: Adalbert Lazăr <alazar@bitdefender.com>