reconnect when the introspector is not ready
QEMU starts a client connection to the introspector and stops if it
cannot connect. This is the "-chardev socket,..." part, before the start
of the introspection handshake (vm_introspection_connect()). Adding
',reconnect=N' (N>0) to the chardev object will let the code flow
reach the vm_introspection_connect() function, which will connect to
the introspector or start a thread to do it when the socket connection
is done.

However, when using the reconnect argument while the introspector is
ready, by the time vm_introspection_connect() is executed, the socket
won't be ready and the connection to the introspector will be delayed
with at least 2 seconds and with an warning.

Another ugly thing is that a sleep(2) is used when waiting for the
socket connection. The qemu_chr_fe_wait_connected() function doesn't
wait because the object used for waiting is not yet created.

Besides this, the patch does a couple of other things:
 - moves the introspection code from kvm-all.c to vm-introspection.c
 - uses better error reporting
 - apply some code refactoring

Signed-off-by: Adalbert Lazăr <alazar@bitdefender.com>