kvm: vmi: store/restore 'vm_start_time' on migrate/snapshot
The VM start time sent during handshake can be used by the introspection
tool as a session id.

We ensure that this 'VM start time' is saved when a snapshot is made
and send it again to the introspection tool when the VM is restored
from snapshot and the introspection connection is reestablished.

Signed-off-by: Adalbert Lazăr <alazar@bitdefender.com>