kvm: vmi: block the object destruction if the chardev is connect
The introspection tool can modify the VM while it is running
(e.g. it can set breakpoints), and when the VM is no longer introspected
these changes need to be removed. Until then, we block the destruction of
the object that would lead to the unexpected shutdown of the introspection
channel.

Signed-off-by: Adalbert Lazăr <alazar@bitdefender.com>